+++ mm2src/coins/nft_storage/mod.rs
@@ -39,7 +39,7 @@ pub trait NftListStorageOps {
         page_number: Option<NonZeroUsize>,
     ) -> MmResult<NftList, Self::Error>;
 
-    async fn add_nfts_to_list<I>(&self, chain: &Chain, nfts: I, last_scanned_block: u32) -> MmResult<(), Self::Error>
+    async fn add_nfts_to_list<I>(&self, chain: &Chain, nfts: I, last_scanned_block: u64) -> MmResult<(), Self::Error>
     where
         I: IntoIterator<Item = Nft> + Send + 'static,
         I::IntoIter: Send;
@@ -69,11 +69,11 @@ pub trait NftListStorageOps {
     async fn refresh_nft_metadata(&self, chain: &Chain, nft: Nft) -> MmResult<(), Self::Error>;
 
     /// `get_last_block_number` function returns the height of last block in NFT LIST table
-    async fn get_last_block_number(&self, chain: &Chain) -> MmResult<Option<u32>, Self::Error>;
+    async fn get_last_block_number(&self, chain: &Chain) -> MmResult<Option<u64>, Self::Error>;
 
     /// `get_last_scanned_block` function returns the height of last scanned block
     /// when token was added or removed from MFT LIST table.
-    async fn get_last_scanned_block(&self, chain: &Chain) -> MmResult<Option<u32>, Self::Error>;
+    async fn get_last_scanned_block(&self, chain: &Chain) -> MmResult<Option<u64>, Self::Error>;
 
     /// `update_nft_amount` function sets a new amount of a particular token in NFT LIST table
     async fn update_nft_amount(&self, chain: &Chain, nft: Nft, scanned_block: u64) -> MmResult<(), Self::Error>;
@@ -105,14 +105,14 @@ pub trait NftTxHistoryStorageOps {
         I: IntoIterator<Item = NftTransferHistory> + Send + 'static,
         I::IntoIter: Send;
 
-    async fn get_last_block_number(&self, chain: &Chain) -> MmResult<Option<u32>, Self::Error>;
+    async fn get_last_block_number(&self, chain: &Chain) -> MmResult<Option<u64>, Self::Error>;
 
     /// `get_txs_from_block` function returns transfers sorted by
     /// block_number in ascending order. It is needed to update the NFT LIST table correctly.
     async fn get_txs_from_block(
         &self,
         chain: &Chain,
-        from_block: u32,
+        from_block: u64,
     ) -> MmResult<Vec<NftTransferHistory>, Self::Error>;
 
     async fn get_txs_by_token_addr_id(
